git-svn-id: https://svn.pjsip.org/repos/pjproject/branches/projects/uwp@5513 74dad513-b988-da41-8d7b-12977e46ad98

+++ pjsip/src/pjsip/sip_dialog.c
@@ -92,6 +92,12 @@ static pj_status_t create_dialog( pjsip_user_agent *ua,
     pj_list_init(&dlg->inv_hdr);
     pj_list_init(&dlg->rem_cap_hdr);
 
+    /* Init client authentication session. */
+    status = pjsip_auth_clt_init(&dlg->auth_sess, dlg->endpt,
+				 dlg->pool, 0);
+    if (status != PJ_SUCCESS)
+	goto on_error;
+
     status = pj_mutex_create_recursive(pool, dlg->obj_name, &dlg->mutex_);
     if (status != PJ_SUCCESS)
 	goto on_error;
@@ -119,6 +125,7 @@ static void destroy_dialog( pjsip_dialog *dlg, pj_bool_t unlock_mutex )
 	pjsip_tpselector_dec_ref(&dlg->tp_sel);
 	pj_bzero(&dlg->tp_sel, sizeof(pjsip_tpselector));
     }
+    pjsip_auth_clt_deinit(&dlg->auth_sess);
     pjsip_endpt_release_pool(dlg->endpt, dlg->pool);
 }
 
@@ -282,12 +289,6 @@ PJ_DEF(pj_status_t) pjsip_dlg_create_uac( pjsip_user_agent *ua,
     /* Initial route set is empty. */
     pj_list_init(&dlg->route_set);
 
-    /* Init client authentication session. */
-    status = pjsip_auth_clt_init(&dlg->auth_sess, dlg->endpt,
-				 dlg->pool, 0);
-    if (status != PJ_SUCCESS)
-	goto on_error;
-
     /* Register this dialog to user agent. */
     status = pjsip_ua_register_dlg( ua, dlg );
     if (status != PJ_SUCCESS)
@@ -311,10 +312,11 @@ on_error:
 /*
  * Create UAS dialog.
  */
-PJ_DEF(pj_status_t) pjsip_dlg_create_uas(   pjsip_user_agent *ua,
-					    pjsip_rx_data *rdata,
-					    const pj_str_t *contact,
-					    pjsip_dialog **p_dlg)
+pj_status_t create_uas_dialog( pjsip_user_agent *ua,
+			       pjsip_rx_data *rdata,
+			       const pj_str_t *contact,
+			       pj_bool_t inc_lock,
+			       pjsip_dialog **p_dlg)
 {
     pj_status_t status;
     pjsip_hdr *pos = NULL;
@@ -504,11 +506,11 @@ PJ_DEF(pj_status_t) pjsip_dlg_create_uas(   pjsip_user_agent *ua,
     }
     dlg->route_set_frozen = PJ_TRUE;
 
-    /* Init client authentication session. */
-    status = pjsip_auth_clt_init(&dlg->auth_sess, dlg->endpt,
-				 dlg->pool, 0);
-    if (status != PJ_SUCCESS)
-	goto on_error;
+    /* Increment the dialog's lock since tsx may cause the dialog to be
+     * destroyed prematurely (such as in case of transport error).
+     */
+    if (inc_lock)
+        pjsip_dlg_inc_lock(dlg);
 
     /* Create UAS transaction for this request. */
     status = pjsip_tsx_create_uas(dlg->ua, rdata, &tsx);
@@ -552,8 +554,40 @@ on_error:
 	--dlg->tsx_count;
     }
 
-    destroy_dialog(dlg, PJ_FALSE);
+    if (inc_lock) {
+        pjsip_dlg_dec_lock(dlg);
+    } else {
+        destroy_dialog(dlg, PJ_FALSE);
+    }
+    
     return status;
+}
+
+
+#if !DEPRECATED_FOR_TICKET_1902
+/*
+ * Create UAS dialog.
+ */
+PJ_DEF(pj_status_t) pjsip_dlg_create_uas(   pjsip_user_agent *ua,
+					    pjsip_rx_data *rdata,
+					    const pj_str_t *contact,
+					    pjsip_dialog **p_dlg)
+{
+    return create_uas_dialog(ua, rdata, contact, PJ_FALSE, p_dlg);
+}
+#endif
+
+
+/*
+ * Create UAS dialog and increase its session count.
+ */
+PJ_DEF(pj_status_t)
+pjsip_dlg_create_uas_and_inc_lock(    pjsip_user_agent *ua,
+				      pjsip_rx_data *rdata,
+				      const pj_str_t *contact,
+				      pjsip_dialog **p_dlg)
+{
+    return create_uas_dialog(ua, rdata, contact, PJ_TRUE, p_dlg);
 }
 
 
@@ -748,11 +782,13 @@ static pj_status_t unregister_and_destroy_dialog( pjsip_dialog *dlg,
     /* MUST not have pending transactions. */
     PJ_ASSERT_RETURN(dlg->tsx_count==0, PJ_EINVALIDOP);
 
-    /* Unregister from user agent. */
-    status = pjsip_ua_unregister_dlg(dlg->ua, dlg);
-    if (status != PJ_SUCCESS) {
-	pj_assert(!"Unexpected failed unregistration!");
-	return status;
+    /* Unregister from user agent, if it has been registered (see #1924) */
+    if (dlg->dlg_set) {
+	status = pjsip_ua_unregister_dlg(dlg->ua, dlg);
+	if (status != PJ_SUCCESS) {
+	    pj_assert(!"Unexpected failed unregistration!");
+	    return status;
+	}
     }
 
     /* Log */
@@ -1088,6 +1124,9 @@ static pj_status_t dlg_create_request_throw( pjsip_dialog *dlg,
     if (status != PJ_SUCCESS)
 	return status;
 
+    /* Put this dialog in tdata's mod_data */
+    tdata->mod_data[dlg->ua->id] = dlg;
+
     /* Just copy dialog route-set to Route header.
      * The transaction will do the processing as specified in Section 12.2.1
      * of RFC 3261 in function tsx_process_route() in sip_transaction.c.
@@ -1187,6 +1226,9 @@ PJ_DEF(pj_status_t) pjsip_dlg_send_request( pjsip_dialog *dlg,
 
     /* Lock and increment session */
     pjsip_dlg_inc_lock(dlg);
+
+    /* Put this dialog in tdata's mod_data */
+    tdata->mod_data[dlg->ua->id] = dlg;
 
     /* If via_addr is set, use this address for the Via header. */
     if (dlg->via_addr.host.slen > 0) {
@@ -1369,6 +1411,9 @@ PJ_DEF(pj_status_t) pjsip_dlg_create_response(	pjsip_dialog *dlg,
 
     /* Lock the dialog. */
     pjsip_dlg_inc_lock(dlg);
+
+    /* Put this dialog in tdata's mod_data */
+    tdata->mod_data[dlg->ua->id] = dlg;
 
     dlg_beautify_response(dlg, PJ_FALSE, st_code, tdata);
